pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"


[project]
version = "0.0.0"
name = "herold"

requires-python = ">=3.8"

dependencies = [
  "pyyaml",
  "jinja2",
  "simple-logging-setup",
]


[project.optional-dependencies]
dev = [
  "build",
  "twine",
  "tox",
  "rlpython",
]

test = [
  "coverage==7.4.4",
  "pytest==8.1.1",
]


[tool.setuptools]
include-package-data = true

script-files = [
  "bin/herold",
]

[tool.setuptools.packages.find]
include = ["herold*"]

[tool.setuptools.package-data]
"herold.default_config" = ["**"]
"herold.hooks" = ["**"]

[tool.setuptools.exclude-package-data]
"*" = [
  ".*",
  "~*",
  "*.swp",
  "*.pyc",
]
